Fix: several bugs exceeded the upper limit of int when calculating large systems 	1. Local_Orbital_Charge::allocate_gamma() 	2. HamiltLCAO::HamiltLCAO() 	3. ELPA_Solver::allocate_work() 	4. ELPA_Solver::decomposeRightMatrix() 	5. ELPA_Solver::verify() 	6. Psi::resize() 	7. Psi::size()

+++ source/module_hsolver/genelpa/elpa_new.cpp
@@ -432,7 +432,7 @@ int ELPA_Solver::read_complex_kernel()
 
 int ELPA_Solver::allocate_work()
 {
-    unsigned long nloc = narows * nacols; // local size
+    unsigned long nloc = static_cast<unsigned long>(narows) * nacols; // local size
     unsigned long maxloc; // maximum local size
     MPI_Allreduce(&nloc, &maxloc, 1, MPI_UNSIGNED_LONG, MPI_MAX, comm);
     maxloc = nloc;
